move `heatmap` to dsi (not under offline)
